fixing use of unsigned int for strlen output
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@4014 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/encodingAscii.c
@@ -82,7 +82,7 @@ _nrrdEncodingAscii_write(FILE *file, const void *_data, size_t elNum,
                          const Nrrd *nrrd, NrrdIoState *nio) {
   char me[]="_nrrdEncodingAscii_write", err[BIFF_STRLEN],
     buff[AIR_STRLEN_MED];
-  unsigned int bufflen, linelen;
+  size_t bufflen, linelen;
   const char *data;
   size_t I;
